:sparkles: Implement Tokenizer
