+++ conf/default.conf
@@ -4,13 +4,43 @@
 # Default:	—
 # Context:	main
 http {
+	# Sets configuration for a virtual server.
+	# Syntax:	server { ... }
+	# Default:	—
+	# Context:	http
+	# http://nginx.org/en/docs/http/ngx_http_core_module.html#server
+	# There is no clear separation between IP-based (based on the IP address) and name-based (based on the “Host” request header field) virtual servers. Instead, the listen directives describe all addresses and ports that should accept connections for the server, and the server_name directive lists all server names.
 	server {
-		listen       8181;
-		server_name  webserv1;
+		# Sets the address and port for IP, or the path for a UNIX-domain socket on which the server will accept requests.
+		# Syntax:	listen address[:port] [default_server] [ssl] [http2 | quic] [proxy_protocol] [setfib=number] [fastopen=number] [backlog=number] [rcvbuf=size] [sndbuf=size] [accept_filter=filter] [deferred] [bind] [ipv6only=on|off] [reuseport] [so_keepalive=on|off|[keepidle]:[keepintvl]:[keepcnt]];
+		# listen port [default_server] [ssl] [http2 | quic] [proxy_protocol] [setfib=number] [fastopen=number] [backlog=number] [rcvbuf=size] [sndbuf=size] [accept_filter=filter] [deferred] [bind] [ipv6only=on|off] [reuseport] [so_keepalive=on|off|[keepidle]:[keepintvl]:[keepcnt]];
+		# listen unix:path [default_server] [ssl] [http2 | quic] [proxy_protocol] [backlog=number] [rcvbuf=size] [sndbuf=size] [accept_filter=filter] [deferred] [bind] [so_keepalive=on|off|[keepidle]:[keepintvl]:[keepcnt]];
+		# Default:	listen *:80 | *:8000;
+		# Context:	server
+		# http://nginx.org/en/docs/http/ngx_http_core_module.html#listen
+		listen		8181;
+		listen		192.168.1.1:8181;
 
+		# Sets names of a virtual server.
+		# Syntax:	server_name name ...;
+		# Default:	server_name "";
+		# Context:	server
+		# http://nginx.org/en/docs/http/ngx_http_core_module.html#server_name
+		# Memo: 1. server_name can be multiple names separated by spaces.
+		#       2. server_name directive can be used multiple times.
+		server_name  webserv1;
+		server_name	example.com www.example.com;
+
+		# Sets configuration depending on a request URI.
+		# Syntax:	location [ = | ~ | ~* | ^~ ] uri { ... }
+		# 			location @name { ... }
+		# Default:	—
+		# Context:	server, location
+		# https://nginx.org/en/docs/http/ngx_http_core_module.html#location
 		location / {
 			root ./html;
 		}
+
 		# Sets the root directory for requests
 		# Syntax:	root path;
 		# Default:	root html;
@@ -63,12 +93,13 @@ http {
 		# Context:	location
 		# https://nginx.org/en/docs/http/ngx_http_core_module.html#limit_except
 		location /limit_except/ {
-			root ./html
+			root ./html;
 			# Allow GET and POST
-			limit_except GET POST {
-				allow 192.168.1.0/32;
-				deny  all;
-			}
+			limit_except GET POST;
+			#limit_except GET POST {
+			#	allow 192.168.1.0/32;
+			#	deny  all;
+			#}
 		}
 
 		# Enables or disables the directory listing output.
@@ -108,12 +139,12 @@ http {
 		# Context:	http, server, location
 		# https://nginx.org/en/docs/http/ngx_http_core_module.html#client_max_body_size
 		location /client_max_body_size/ {
-			root ./html
+			root ./html;
 			client_max_body_size 1m;
 		}
 		# Only file which size is less than 1MB can be uploaded.
 		location /upload/limit {
-			upload_store ./html/upload
+			upload_store ./html/upload;
 			client_max_body_size 1m;
 		}
 
@@ -136,43 +167,6 @@ http {
 		location / {
 			root	./html/webserv2;
 			index	hello.html;
-		}
-	}
-}
-
-http {
-	# Sets configuration for a virtual server.
-	# Syntax:	server { ... }
-	# Default:	—
-	# Context:	http
-	# http://nginx.org/en/docs/http/ngx_http_core_module.html#server
-	# There is no clear separation between IP-based (based on the IP address) and name-based (based on the “Host” request header field) virtual servers. Instead, the listen directives describe all addresses and ports that should accept connections for the server, and the server_name directive lists all server names.
-	server {
-		# Sets the address and port for IP, or the path for a UNIX-domain socket on which the server will accept requests.
-		# Syntax:	listen address[:port] [default_server] [ssl] [http2 | quic] [proxy_protocol] [setfib=number] [fastopen=number] [backlog=number] [rcvbuf=size] [sndbuf=size] [accept_filter=filter] [deferred] [bind] [ipv6only=on|off] [reuseport] [so_keepalive=on|off|[keepidle]:[keepintvl]:[keepcnt]];
-		# listen port [default_server] [ssl] [http2 | quic] [proxy_protocol] [setfib=number] [fastopen=number] [backlog=number] [rcvbuf=size] [sndbuf=size] [accept_filter=filter] [deferred] [bind] [ipv6only=on|off] [reuseport] [so_keepalive=on|off|[keepidle]:[keepintvl]:[keepcnt]];
-		# listen unix:path [default_server] [ssl] [http2 | quic] [proxy_protocol] [backlog=number] [rcvbuf=size] [sndbuf=size] [accept_filter=filter] [deferred] [bind] [so_keepalive=on|off|[keepidle]:[keepintvl]:[keepcnt]];
-		# Default:	listen *:80 | *:8000;
-		# Context:	server
-		# http://nginx.org/en/docs/http/ngx_http_core_module.html#listen
-		listen      192.168.1.1:8181;
-
-		# Sets names of a virtual server.
-		# Syntax:	server_name name ...;
-		# Default:	server_name "";
-		# Context:	server
-		# http://nginx.org/en/docs/http/ngx_http_core_module.html#server_name
-		server_name	example.com www.example.com;
-
-		# Sets configuration depending on a request URI.
-		# Syntax:	location [ = | ~ | ~* | ^~ ] uri { ... }
-		# 			location @name { ... }
-		# Default:	—
-		# Context:	server, location
-		# https://nginx.org/en/docs/http/ngx_http_core_module.html#location
-		location / {
-			root	./html;
-			index	world.html;
 		}
 	}
 }
